Add double-click card to show card feature; rename `PackFrame` to `PackList` frame and `CardFrame` to `CardList` frame.
